some changes

+++ src/Dot.java
@@ -1,25 +1,25 @@
 public class Dot {
-    private int mX;
-    private int mY;
+    private long mX;
+    private long mY;
 
-    public void setX(int x) {
+    public void setX(long x) {
         mX = x;
     }
 
-    public void setY(int y) {
+    public void setY(long y) {
         mY = y;
     }
 
-    Dot(int x, int y){
+    Dot(long x, long y){
         mX = x;
         mY = y;
     }
 
-    public int getX() {
+    public long getX() {
         return mX;
     }
 
-    public int getY() {
+    public long getY() {
         return mY;
     }
 }
